requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
deepface>=0.0.92
opencv-python>=4.9
numpy>=1.24
pillow>=10.0
python-multipart>=0.0.9
aiofiles>=23.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9
qrcode>=7.4

# Advanced real-time optimization
faiss-cpu>=1.7.4
websockets>=12.0
asyncio-mqtt>=0.16.1

# GPU acceleration (optional)
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3

# Security enhancements
slowapi>=0.1.9
redis>=4.5.0
jinja2>=3.1.0
